Implemented JOIN_TABLE strategy + FIX TextualFact
- TextualFact FIX: class constructor now set textual value.

- JOIN_TABLE strategy: seems that annotations + loading orm-joinTable.xml will cause hibernate to generate the DISCRIMINATOR_COLUMN.. ugly :(

The problem is the annotation in the superclass: @DiscriminatorColumn(name = "***", discriminatorType = DiscriminatorType.STRING )

The annotations in the subclass that define the discriminator value (@DiscriminatorValue("***") ) are ignored if the super class doesn't have the @DiscriminatorColumn annotation.

So, to make an easy and fast switch, the best approach I have found is to define another orm-singleTable.xml to define the superclass inheritance properties instead of using annotation.

Obviously both orm-joinTable.xml and orm-singleTable.xml only specifies the inheritance strategy, the other JPA parameters/mapping settings are defined with annotations in each class.

To chose the strategy, open persistence.xml and comment/uncomment the two lines that chose the orm-***.xml mapping file.
NB: do not select both orm.xml files, only one at the same time.
NB: one of the two files must be selected.