PersistencyHelper + Double DB Management (single/join table)
PersistencyHelper: new class, we can access to all the objects and methods related to persistency that we need in our persistency tests through objects of this class.
We control also the inheritance strategy (and so the DB name):
 - through the PersistencyHelper constructor we can specify the strategy to use for the new connection.
 - through a constant in the PersistencyHelper we can chose the default inheritance strategy for the default constructor (that doesn't require a strategy specification but use the default). This is useful because changing a single line of code we can control all the connection in all tests! (if in that tests we used the default constructors)

Double DB Management: to make possible this configuration switch, we use two different persistency unit in the persistence.xml, that uses two different schemas in the DB and loads the two different orm-singleTable.xml and orm-joinTable.xml, changing the inheritance behavior for Fact and FactType hierarchies.