Implemented Ancestor-List Strategy for both the hierarchies (Fact/Types).
Modified also the persistence.xml to have different DB names with this Anchestor List Strategy rispect to the master branch without this strategy.

NB: to make a fair performance test, it would be good to use a fixed hierarchy for both Types and Facts.
The use of the RandomGenerator is not a good idea because it will generate differents hierarchies with differents depth and width of the tree, with a different number of units, phenomenon, types, and so on..

There are two solutions:
A) Use a fixed pseudo-random generator (feed the pseudo random generator with the same feed so that they generate the same pseudo-random sequence at each execution).

B) Use fixed depth, fixed width, fixed number of phenomenons/units, fixed string length for the Type names and Text Fact values (and for phenomenons values). BTW we can not generate identicals tree because the type of each leaf is chosen randomly (Textual/Quantitative/Qualitative and of course also Composite in the middle levels) etc.. but it's a good solution if at least we force that the leaf are only at the last depth level (NB: we will produce only perfect balanced tree, scenario in which Ancestor-List strategy performs optimally... not so fair performance test!).