Ancestor Strategy fix
- Ancestor association is ManyToMany not OneToMany.
- Added new procedures to fix a problem with the ancestors:
If i have 4 elements A, B, C, D and if I link A with B, then C with D...
A-->B
C-->D

...I can say:
- Ancestors-list of A = {}
- Ancestors-list of B = {A}
- Ancestors-list of C = {}
- Ancestors-list of D = {C}

Now if i link
B-->C

Using the old procedure, I only add B and all ancestors of A to the ancestors-list of C, but I don't modify the ancestors-list of the C's childs! (in this example, I don't modify the ancestors-list of D)
So I have:

A-->B ---> C--->D
- Ancestors-list of A = {}
- Ancestors-list of B = {A}
- Ancestors-list of C = {B,A}
- Ancestors-list of D = {C}   <-- error, missing A and B.

With the new procedure, I check if the child that I'm adding to the composition (in this case when I add B to C) it's in turn a composition:
if it is, I add recursively to all the descendant of the child (C) all the ancestors that I was adding to C (so I add {A, B} not only to C, but also to all the childs of C, in this example to D).

NB: all these procedures doesn't check for loops. These procedures don't work with loops.