Imlemented Composite/Composed with/without ancesters-list strategy as separate classes
These class are not used for now. Only implemented for future use.

Packages:
- reflectionPattern.utility.composite
- reflectionPattern.utility.compositeWithAncestors

An example on how to use these classes can be found in:

- reflectionPattern.utility.compositeWithAncestors/compositeTest
- reflectionPattern.utility.compositeWithAncestors/compositeTestWithSuperclass

This package (both with and without ancestors) can be used in two different ways:

1) -- MyComponent extends Component --
(example in package compositeTest)

My own class that represent the Component (MyComponent) extends the class Component<MyComposite>.

My own class that represent the Composite (MyComposite), have a private instance (named compositeManager) of the class (with type parameters) CompositeManager<MyComposite, MyComponent>, and implements the interface IComposite<MyComponent>, and so have to implements the methods addChilds(..) and getChilds() simply calling the same methods of the compositeManager.

2) -- MyComponent have an instance of ComponentManager --
(example in package compositeTestWithSuperclass)

My own class that represent the Component (MyComponent) have a private instance (named componentManager) of the class (with type parameters) ComponentManager<MyComponent, MyComposite> and implements the interface IComponent<MyComposite>, and so have to implements the methods addFirstAncestors(), addLastAncestors(), appendAllAncestors(), getAncestors(), getFather() simply calling the same methods of the componentManager.

MyComposite: see the case (1).